change int back to float for gpa

+++ studentList.cpp
@@ -9,7 +9,7 @@ struct Student {
   char firstName[20]; //Struct code based on Mr. Galbraith's video on Structs in Canvas
   char secondName[20];
   int id;
-  int gpa; //a Student has a first name, last name, id, and gpa 
+  float gpa; //a Student has a first name, last name, id, and gpa 
 }
 
 int main() {
